pyproject.toml
[project]
name = "video-to-essay"
version = "0.1.0"
description = "Convert YouTube videos into illustrated markdown essays"
requires-python = ">=3.14"
dependencies = [
    "agentmail>=0.4.0",
    "anthropic>=0.79.0",
    "boto3>=1.38.0",
    "imagehash>=4.3.2",
    "markdown>=3.7",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "posthog>=3.0",
    "sentry-sdk>=2.0",
    "typer>=0.23.0",
    "yt-dlp>=2026.3",
]

[project.scripts]
video-to-essay = "video_to_essay.main:app"

[dependency-groups]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/video_to_essay"]
